[PPANTT-41] feat: Hours summary section & clean code
